The Perfect Guess
This program generates a random number and asks the user to guess it.
If the player's guess is higher than the actual number, the program displays "Lower number please". Similarly if the user's guess is too low, the program displays the number of guesses the player used to arrive at the number.